Added Python 3 support.  This required renaming unicode.py to unicode_chars.py (since the 2to3 process likes to rename all insances of 'unicode' with 'str') and a few changes to setup.py.  Also, the odict module will now only be listed as a requirement if using Python 2.6 or older (it was added to the standard library in Python 2.7).
